Add styles with settings
I tried to model styles as per ITCSS as much as I could, but I'll be
using styled-components, which will act as tools & components.

Also, settings are overly complex, this is because I maintain my
settings preferences, will not hinger performance as it will be clean-up
with webpack's tree shake.